Add test coverage, fix clippy warnings, add lint targets (Stage 5)
- Fix all 8 clippy warnings (needless borrows, matches! macro,
  unnecessary casts, let-and-return, filter-next, map-clone)
- Add Makefile targets: lint, fmt, fmt-fix, check (lint+fmt+test)
- Add full proxy lifecycle test (insert→verify→circuit breaker→recovery)
- Add collect_top_n correctness test with 50 proxies
- Add config hot-reload test (write→load→rewrite→reload→verify)
- Add config malformed JSON fallback test
- Add API tests: add_proxy, add_proxy_duplicate, ban invalid key,
  ban invalid port, ban+unban cycle, prometheus format, health with proxies

Tests: 85 passing (was 74), 0 clippy warnings

+++ src/pool/metrics.rs
@@ -59,7 +59,7 @@ pub fn collect(state: &SharedState) -> Metrics {
     }
 
     let (total_successes, total_fails) = proxies.iter().fold((0u64, 0u64), |(s, f), p| {
-        (s + p.success_count as u64, f + p.fail_count as u64)
+        (s + p.success_count, f + p.fail_count)
     });
 
     let success_rate = if total_successes + total_fails > 0 {
